Reading and writting in chunks
For GET requets, instead of reading the whole file and then
sending it, read slices of 1 KiB. This allows us to send bigger files.
For POST requests, instead of reading the data of the request inside the
Request::parse method, add a function to read it on demand. This saves
time in case the request data is not needed (e.g. if the url is invalid)